% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaunay.R
\name{cdel}
\alias{cdel}
\title{Title}
\usage{
cdel(points, edges)
}
\arguments{
\item{points}{xx}

\item{edges}{xx}
}
\value{
xx
}
\description{
Title
}
\examples{
nsides <- 12L
angles <- seq(0, 2*pi, length.out = nsides+1L)[-1L]
points <- cbind(cos(angles), sin(angles))
points <- rbind(points, points/1.5)
# constraint edges
indices <- 1L:nsides
edges_outer <- cbind(
  indices, c(indices[-1L], indices[1L])
)
edges_inner <- edges_outer + nsides
edges <- rbind(edges_outer, edges_inner)
cdel(points, edges)
}
