% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{plotVoronoi}
\alias{plotVoronoi}
\title{Plot Voronoï diagram}
\usage{
plotVoronoi(
  tessellation,
  colors = "random",
  hue = "random",
  luminosity = "light",
  alpha = 1,
  ...
)
}
\arguments{
\item{tessellation}{an output of \code{\link{Voronoi}}}

\item{colors}{this can be \code{"random"} to use random colors for the cells
(with \code{\link[randomcoloR]{randomColor}}), \code{"distinct"} to use
distinct colors with the help of
\code{\link[randomcoloR]{distinctColorPalette}}, \code{NA} for no colors, 
or this can be a vector of colors; the length of this vector
of colors must match the number of Voronoï cells, that you can get by
typing \code{length(tessellation)}}

\item{hue, luminosity}{if \code{colors = "random"}, these arguments are passed
to \code{\link[randomcoloR]{randomColor}}}

\item{alpha}{opacity, a number between 0 and 1
(used when \code{colors} is not \code{NA})}

\item{...}{arguments passed to \code{\link[graphics]{polygon}} to plot the 
cells}
}
\value{
No returned value.
}
\description{
Plot a Voronoï tessellation.
}
\examples{
library(delaunay)
}
